wxGUI/wxplot: colors should be integers
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@74186 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ gui/wxpython/wxplot/base.py
@@ -132,9 +132,9 @@ class BasePlotFrame(wx.Frame):
         for clr in six.iterkeys(grass.named_colors):
             if clr == 'white':
                 continue
-            r = grass.named_colors[clr][0] * 255
-            g = grass.named_colors[clr][1] * 255
-            b = grass.named_colors[clr][2] * 255
+            r = int(grass.named_colors[clr][0] * 255)
+            g = int(grass.named_colors[clr][1] * 255)
+            b = int(grass.named_colors[clr][2] * 255)
             self.colorDict[clr] = (r, g, b, 255)
 
     def InitPlotOpts(self, plottype):
